// frontends/epro/src/app/api/model/event-dto.ts
/**
 * KV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: DEV-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventModelDTO } from './event-model-dto';
import { WorkflowActionDTO } from './workflow-action-dto';
import { WorkflowStatusDTO } from './workflow-status-dto';


export interface EventDTO { 
    scopePk: number;
    scopeId: string;
    scopeCodeAndShortname: string;
    pk: number;
    id: string;
    creationTime: Date;
    lastUpdateTime: Date;
    shortname: string;
    longname: string;
    model: EventModelDTO;
    modelId: string;
    eventGroupNumber: number;
    removed: boolean;
    inRemoved: boolean;
    canWrite: boolean;
    canBeRemoved: boolean;
    expectedDate: Date;
    date: Date;
    endDate?: Date;
    mandatory: boolean;
    expected: boolean;
    notDone: boolean;
    blocking: boolean;
    locked: boolean;
    workflowStatus: Array<WorkflowStatusDTO>;
    workflowCreationActions: Array<WorkflowActionDTO>;
}